refactor: member 패키지 -> auth 패키지 이동
