+++ src/main/java/com/woowa/woowakit/domain/auth/domain/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.woowa.woowakit.domain.auth.domain;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByEmail(Email email);
+}
