+++ src/main/java/com/woowa/woowakit/domain/auth/domain/PasswordEncoder.java
@@ -1,0 +1,8 @@
+package com.woowa.woowakit.domain.auth.domain;
+
+public interface PasswordEncoder {
+
+    boolean matches(String planePassword, String encodedPassword);
+
+    String encode(String password);
+}
