+++ src/main/java/com/woowa/woowakit/domain/auth/domain/Role.java
@@ -1,0 +1,8 @@
+package com.woowa.woowakit.domain.auth.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum Role {
+    ADMIN, USER
+}
