+++ src/main/java/com/woowa/woowakit/domain/auth/dto/request/LoginRequest.java
@@ -1,0 +1,25 @@
+package com.woowa.woowakit.domain.auth.dto.request;
+
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+public class LoginRequest {
+
+    @Email
+    private String email;
+
+    @NotBlank
+    private String password;
+
+    public static LoginRequest of(String email, String password) {
+        return new LoginRequest(email, password);
+    }
+}
