+++ src/main/java/com/woowa/woowakit/domain/auth/dto/request/SignUpRequest.java
@@ -1,0 +1,26 @@
+package com.woowa.woowakit.domain.auth.dto.request;
+
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+public class SignUpRequest {
+
+    @Email
+    private String email;
+    @NotBlank
+    private String password;
+    @NotBlank
+    private String name;
+
+    public static SignUpRequest of(String email, String password, String name) {
+        return new SignUpRequest(email, password, name);
+    }
+}
