+++ src/main/java/com/woowa/woowakit/domain/auth/exception/EmailInvalidException.java
@@ -1,0 +1,10 @@
+package com.woowa.woowakit.domain.auth.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class EmailInvalidException extends MemberException {
+
+    public EmailInvalidException() {
+        super("이메일 형식이 올바르지 않습니다.", HttpStatus.BAD_REQUEST);
+    }
+}
