+++ src/main/java/com/woowa/woowakit/domain/auth/exception/MemberException.java
@@ -1,0 +1,27 @@
+package com.woowa.woowakit.domain.auth.exception;
+
+import com.woowa.woowakit.global.error.WooWaException;
+import org.springframework.http.HttpStatus;
+
+public class MemberException extends WooWaException {
+
+    public MemberException(HttpStatus httpStatus) {
+        super(httpStatus);
+    }
+
+    public MemberException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public MemberException(String message, Throwable cause, HttpStatus httpStatus) {
+        super(message, cause, httpStatus);
+    }
+
+    public MemberException(String message) {
+        super(message);
+    }
+
+    public MemberException(String message, HttpStatus httpStatus) {
+        super(message, httpStatus);
+    }
+}
