--- src/main/java/com/woowa/woowakit/domain/member/domain/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.woowa.woowakit.domain.member.domain;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(Email email);
-}
