--- src/main/java/com/woowa/woowakit/domain/member/domain/PasswordEncoder.java
@@ -1,8 +1,0 @@
-package com.woowa.woowakit.domain.member.domain;
-
-public interface PasswordEncoder {
-
-    boolean matches(String planePassword, String encodedPassword);
-
-    String encode(String password);
-}
