--- src/main/java/com/woowa/woowakit/domain/member/domain/Role.java
@@ -1,8 +1,0 @@
-package com.woowa.woowakit.domain.member.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum Role {
-    ADMIN, USER
-}
