--- src/main/java/com/woowa/woowakit/domain/member/dto/response/LoginResponse.java
@@ -1,18 +1,0 @@
-package com.woowa.woowakit.domain.member.dto.response;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public class LoginResponse {
-
-    private String accessToken;
-
-    public static LoginResponse from(String accessToken) {
-        return new LoginResponse(accessToken);
-    }
-}
