--- src/main/java/com/woowa/woowakit/domain/member/exception/LoginFailException.java
@@ -1,10 +1,0 @@
-package com.woowa.woowakit.domain.member.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class LoginFailException extends MemberException {
-
-    public LoginFailException() {
-        super("이메일 혹은 비밀번호를 잘못 입력하셨습니다.", HttpStatus.BAD_REQUEST);
-    }
-}
